#63 register and login for client user

+++ carRental/src/main/java/com/amigoscode/chohort2/carRental/driverLicense/DriverLicense.java
@@ -36,7 +36,7 @@ public class DriverLicense extends AbstractAuditingEntity {
     private String driverLicenseNumber;
 
     @Column (name = "issued_in", nullable = false)
-    private Integer issuedIn;
+    private Long issuedIn;
 
     @Column (name = "issued_at", nullable = false)
     private LocalDate issuedAt;
